Updated PokemonManagerBase to play nicer with !psearch items that required a user parameter. Fixed a crash by adding in abbreviations to Version constructor. Fleshed out Pokemon !psearch fields to the current limits of the Pokemon object.

+++ src/includes/modules/Pokemon/Stats/ParameterParser.php
@@ -375,9 +375,9 @@ class IVStatParameterResult {
 
     /**
      * @param mixed $stat Stat name or index
-     * @return int Nature multiplier value, or -1 on failure
+     * @return float Nature multiplier value, or -1 on failure
      */
-    public function getNatureMultiplier($stat): int {
+    public function getNatureMultiplier($stat): float {
         $key = $this->getKey($stat);
 
         return $this->natureMultipliers[ $key ] ?? $key;
